backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=23.5,<24.0
django-cors-headers>=4.3,<5.0
django-environ>=0.11,<1.0
django-redis>=5.4,<6.0
drf-spectacular>=0.27,<1.0
django-ratelimit>=4.1,<5.0
psycopg[binary]>=3.1,<4.0
Pillow>=10.0,<11.0
markdown>=3.5,<4.0
django-jazzmin>=3.0,<4.0
whitenoise>=6.6,<7.0
